+++ apps/ranger-next/pages/index.tsx
@@ -1,11 +1,17 @@
+import { useEffect } from 'react'
 import { fetchEdege, HeadElement, HtmlELement, ScriptElement } from '@ranger-theme/adobe-edege'
 
-const url: string = 'https://main--aem-block-collection--adobe.hlx.live'
+const url: string = process.env.NEXT_PUBLIC_EDEGE_URL
+const host: string = process.env.NEXT_PUBLIC_HOST_URL
 
 const Home = ({ html }: { html: string }) => {
+  useEffect(() => {
+    window?.edegeLoadPage?.()
+  }, [])
+
   return (
     <div>
-      <HeadElement html={html} />
+      <HeadElement host={host} html={html} url={url} />
       <ScriptElement html={html} url={url} />
       <div>
         <HtmlELement html={html} url={url} />