+++ packages/ranger-adobe-edege/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ranger-theme/adobe-edege
 
+## 0.4.0
+
+### Minor Changes
+
+- edege router change reload events
+
+## 0.3.0
+
+### Minor Changes
+
+- support head meta components
+
 ## 0.2.0
 
 ### Minor Changes